Expose html to block converter through Draft object
